+++ libs/util-matrix/src/index.ts
@@ -10,6 +10,7 @@ export * from './lib/poll';
 export * from './lib/transient-errors';
 export * from './lib/password-uia';
 export * from './lib/attachment-crypto';
+export * from './lib/key-file-crypto';
 export * from './lib/authed-media';
 export * from './lib/room-create';
 export * from './lib/crypto-wasm-loader';